requirements.txt
astroid==2.4.0
awsebcli==3.17.1
bcrypt==3.1.7
beautifulsoup4==4.9.0
blinker==1.4
boto3==1.12.25
botocore==1.14.17
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.3.9
cryptography==2.9
dnspython==1.16.0
docutils==0.15.2
email-validator==1.0.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.16.0
gunicorn==20.0.4
idna==2.7
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
oauthlib==3.1.0
pathspec==0.5.9
Pillow==7.1.2
psycopg2==2.8.4
pycparser==2.20
pylint==2.5.0
pyOpenSSL==19.1.0
pypiwin32==223
python-dateutil==2.8.0
pywin32==227
PyYAML==5.2
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
termcolor==1.1.0
toml==0.10.0
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
